@@ -18,38 +18,38 @@ import AWSLambdaRuntime
 import AWSLambdaTesting
 import Logging
 import NIO
-
-extension Lambda {
-    public static func test<Handler: LambdaHandler>(
-        _ handlerType: Handler.Type,
-        with event: Handler.Event,
-        using config: TestConfig = .init()
-    ) async throws -> Handler.Output {
-        let logger = Logger(label: "test")
-        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-        defer {
-            try! eventLoopGroup.syncShutdownGracefully()
-        }
-        let eventLoop = eventLoopGroup.next()
-
-        let initContext = LambdaInitializationContext.__forTestsOnly(
-            logger: logger,
-            eventLoop: eventLoop
-        )
-
-        let context = LambdaContext.__forTestsOnly(
-            requestID: config.requestID,
-            traceID: config.traceID,
-            invokedFunctionARN: config.invokedFunctionARN,
-            timeout: config.timeout,
-            logger: logger,
-            eventLoop: eventLoop
-        )
-        let handler = try await Handler(context: initContext)
-        defer {
-            let eventLoop = initContext.eventLoop.next()
-            try? initContext.terminator.terminate(eventLoop: eventLoop).wait()
-        }
-        return try await handler.handle(event, context: context)
-    }
-}
+//
+//extension Lambda {
+//    public static func test<Handler: LambdaHandler>(
+//        _ handlerType: Handler.Type,
+//        with event: Handler.Event,
+//        using config: TestConfig = .init()
+//    ) async throws -> Handler.Output {
+//        let logger = Logger(label: "test")
+//        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+//        defer {
+//            try! eventLoopGroup.syncShutdownGracefully()
+//        }
+//        let eventLoop = eventLoopGroup.next()
+//
+//        let initContext = LambdaInitializationContext.__forTestsOnly(
+//            logger: logger,
+//            eventLoop: eventLoop
+//        )
+//
+//        let context = LambdaContext.__forTestsOnly(
+//            requestID: config.requestID,
+//            traceID: config.traceID,
+//            invokedFunctionARN: config.invokedFunctionARN,
+//            timeout: config.timeout,
+//            logger: logger,
+//            eventLoop: eventLoop
+//        )
+//        let handler = try await Handler(context: initContext)
+//        defer {
+//            let eventLoop = initContext.eventLoop.next()
+//            try? initContext.terminator.terminate(eventLoop: eventLoop).wait()
+//        }
+//        return try await handler.handle(event, context: context)
+//    }
+//}
